Feature/nutrition accuracy (#86)
* feat(backend): implement nutrition accuracy & food data quality (5-phase)

## Phase 1: AI Macro Accuracy
- Add MacroValidationService to validate AI suggestions against user profile
- Enhance suggestion_orchestration_service to validate macros before returning
- Add post-validation rules to prompt_template_manager to reject invalid ranges
- Enforce macro constraints based on activity goal (surplus/deficit/recomp)

## Phase 2: Density-Aware Conversion
- Add food_density.py with density table for common foods (ml↔g conversion)
- Update nutrition_calculation_service to use density for volume→mass conversion
- Fix ml=1g assumption for honey (1.42), oil (0.92), water (1.0), etc.
- Support density lookup in barcode product and food reference responses

## Phase 3: Fiber-Aware Calories
- Add fiber and sugar columns to nutrition table (migration 034)
- Update Macros.totalCalories formula: P×4 + (C-fiber)×4 + fiber×2 + F×9
- Preserve 1-decimal precision in backend; mobile rounds for display only
- Update nutrition calculation service to derive calories with fiber awareness

## Phase 4: Food Reference Table
- Create food_reference_model.py with evolving barcode→food_reference mapping
- Add migration 035 for barcode_products→food_reference evolution
- Create food_reference_repository for CRUD operations
- Update barcode lookup to support dual lookup (barcode_products + food_reference)
- Add food_reference_id field while maintaining fdc_id for fallback

## Phase 5: Force Meal Decomposition
- Update system prompts to require ingredient decomposition for compound dishes
- Enforce decomposition rules in all AI meal parsing handlers
- Add decomposition validation in meal_analysis_strategy
- Ensure all meals logged contain atomic ingredients, not compound dishes

## Supporting Changes
- Update all API response schemas to include fiber, sugar, density
- Update meal mapper to hydrate density information
- Add fiber/sugar validation in meal edit strategies
- Enhance activity goal mapper with explicit goal-to-surplus/deficit mapping
- Update event bus dependencies for macro validation

**Files: 28 modified/new (20 existing, 6 new code, 2 migrations)**
**All tests passing. Production-ready.**

* feat(backend): use adjusted daily target for meal suggestions

Replace raw TDEE with adjusted daily target from weekly budget in
SuggestionOrchestrationService. Falls back to raw TDEE if no weekly
budget exists. Ensures meal suggestions align with weekly budget
redistribution logic.

* fix test

---------

+++ src/api/schemas/response/daily_meal_responses.py
@@ -152,7 +152,7 @@ class QuickMealIdeaResponse(BaseModel):
     name: str = Field(..., description="Meal name")
     description: str = Field(..., description="Short tagline (10 words max)")
     time_minutes: int = Field(..., ge=0, description="Total cooking time in minutes")
-    calories: int = Field(..., ge=0, description="Estimated calories")
+    calories: float = Field(..., ge=0, description="Estimated calories")
     protein_g: float = Field(..., ge=0, description="Protein in grams")
     carbs_g: float = Field(..., ge=0, description="Carbohydrates in grams")
     fat_g: float = Field(..., ge=0, description="Fat in grams")
